src/requirements.txt
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.10
babel==2.17.0
binaryornot==0.4.4
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
cryptography==45.0.2
defusedxml==0.7.1
dill==0.4.0
Django==4.2.20
django-appconf==1.1.0
django-browser-reload==1.18.0
django-compressor==4.5.1
django-filter==25.1
django-phonenumber-field==8.1.0
django-tailwind==4.0.1
django-unfold==0.58.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
flake8==7.2.0
gunicorn==20.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.5
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytest-lazy-fixture==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rjsmin==1.2.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
text-unidecode==1.3
tomlkit==0.13.2
types-python-dateutil==2.9.0.20250516
urllib3==2.4.0
whitenoise==6.9.0
